% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasther.R
\name{.LocalHeritability.ini}
\alias{.LocalHeritability.ini}
\title{Initial Heritability Estimation (1 parameter)}
\usage{
.LocalHeritability.ini(Vz2, Uval, n, m, r)
}
\arguments{
\item{Vz2}{Squared projections of Z-statistics}

\item{Uval}{Eigenvalues of LD matrix}

\item{n}{Sample size}

\item{m}{Number of SNPs in genomic region}

\item{r}{Rank after eigenvalue pruning}
}
\value{
List with initial h2 estimate and standard error
}
\description{
Initial Heritability Estimation (1 parameter)
}
\keyword{internal}
