% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasther.R
\name{eigen.Python}
\alias{eigen.Python}
\title{FastHer: Fast and Accurate Local Heritability REML-Estimator using
GWAS summary statistics and LD matrix}
\usage{
eigen.Python(U, only.values = FALSE)
}
\arguments{
\item{U}{A symmetric correlation SNP-by-SNP matrix (LD matrix)}

\item{only.values}{Logical, whether to return only eigenvalues}
}
\value{
List with eigenvalues and eigenvectors
}
\description{
Implements the FastHer method for local heritability estimation
using REML function optimized by eigen-decomposition of LD matrix
Eigen-decomposition Using Python's Scipy
}
\author{
Gulnara R. Svishcheva
}
