% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasther.R
\name{LocalHeritability}
\alias{LocalHeritability}
\title{Core Heritability Estimation Function (2 parameters)}
\usage{
LocalHeritability(Vz2, Uval, n, m, r, start, control_list = NULL)
}
\arguments{
\item{Vz2}{Squared projections of Z-statistics}

\item{Uval}{Eigenvalues of LD matrix}

\item{n}{Sample size}

\item{m}{Number of SNPs}

\item{r}{Rank after eigenvalue pruning}

\item{start}{Initial heritability value}

\item{control_list}{Optimization control parameters (optional)}
}
\value{
List with estimation results
}
\description{
Core Heritability Estimation Function (2 parameters)
}
\keyword{internal}
